Update BSP library and sample code for Keil AC6
1. Add *.uvprojx to the Keil folder of each example.
2. Fix the build problems on Keil V5.28 and V5.38.
3. Fix the build problems on IAR V8.40.
4. Fix the build problems on GCC.

+++ SampleCode/BLE_AB1602/porting/NANO100/bt_porting_spi.c
@@ -137,7 +137,7 @@ void GPDEF_IRQHandler(void)
 void SPITransactionStart(spi_tran_entry *entry)
 {
     if(((uint32_t)entry->txbuf & 0x3) || ((uint32_t)entry->rxbuf & 0x3))
-        printf("spi (%d, %d) = (%x, %x)\n", entry->txlen, entry->rxlen, entry->txbuf, entry->rxbuf);
+        printf("spi (%d, %d) = (%x, %x)\n", entry->txlen, entry->rxlen, (uint32_t)entry->txbuf, (uint32_t)entry->rxbuf);
 
     spi_busy = true;
